+++ PerseusMeteo/BusinessContent/Options/OptionsViewController.swift
@@ -66,16 +66,12 @@ class OptionsViewController: NSViewController, NSTextFieldDelegate {
     @IBOutlet private(set) weak var labelLanguage: NSTextField!
     @IBOutlet private(set) weak var labelTimeFormat: NSTextField!
     @IBOutlet private(set) weak var labelOpenWeatherKey: NSTextField!
-    @IBOutlet private(set) weak var labelStatusMenus: NSTextField!
-    @IBOutlet private(set) weak var labelStatusMenusUpdate: NSTextField!
 
     @IBOutlet private(set) weak var controlDarkMode: NSSegmentedControl!
     @IBOutlet private(set) weak var controlLanguage: NSSegmentedControl!
     @IBOutlet private(set) weak var controlTimeFormat: NSSegmentedControl!
     @IBOutlet private(set) weak var controlOpenWeatherKey: NSTextField!
     @IBOutlet private(set) weak var controlUnlockButton: NSButton!
-    @IBOutlet private(set) weak var checkBoxStatusMenus: NSButton!
-    @IBOutlet private(set) weak var comboBoxStatusMenusUpdatePeriod: NSComboBox!
 
     @IBOutlet private(set) weak var labelTemperature: NSTextField!
     @IBOutlet private(set) weak var labelWindSpeed: NSTextField!
@@ -86,6 +82,21 @@ class OptionsViewController: NSViewController, NSTextFieldDelegate {
     @IBOutlet private(set) weak var controlWindSpeed: NSSegmentedControl!
     @IBOutlet private(set) weak var controlPressure: NSSegmentedControl!
     @IBOutlet private(set) weak var controlDistance: NSSegmentedControl!
+
+    @IBOutlet private(set) weak var labelCurrentWeatherStatusMenus: NSTextField!
+    @IBOutlet private(set) weak var labelStatusMenusUpdate: NSTextField!
+    @IBOutlet private(set) weak var checkBoxCurrentWeatherStatusMenus: NSButton!
+    @IBOutlet private(set) weak var comboBoxStatusMenusUpdatePeriod: NSComboBox!
+
+    @IBOutlet private(set) weak var labelStatusMenus: NSTextField!
+    @IBOutlet private(set) weak var labelSecondLine: NSTextField!
+    @IBOutlet private(set) weak var labelToolTip: NSTextField!
+
+    @IBOutlet private(set) weak var buttonInfo: NSButton!
+    @IBOutlet private(set) weak var checkBoxTwoLines: NSButton!
+    @IBOutlet private(set) weak var comboBoxSecondLine: NSComboBox!
+    @IBOutlet private(set) weak var comboBoxToolTipLeft: NSComboBox!
+    @IBOutlet private(set) weak var comboBoxToolTipRight: NSComboBox!
 
     // MARK: - Actions
 
@@ -133,7 +144,8 @@ class OptionsViewController: NSViewController, NSTextFieldDelegate {
     }
 
     @IBAction func checkBoxStatusMenusDidChanged(_ sender: NSButton) {
-        log.message("[\(type(of: self))].\(#function) - \(checkBoxStatusMenus.state)")
+        let state = checkBoxCurrentWeatherStatusMenus.state
+        log.message("[\(type(of: self))].\(#function) - \(state)")
 
         let isEnabled = sender.state == .on ? true : false
 
@@ -146,6 +158,35 @@ class OptionsViewController: NSViewController, NSTextFieldDelegate {
         log.message("[\(type(of: self))].\(#function) - \(index)")
 
         presenter?.forceStatusMenusUpdatePeriod(sender.indexOfSelectedItem)
+    }
+
+    @IBAction func checkBoxTwoLinesDidChanged(_ sender: NSButton) {
+        let state = checkBoxTwoLines.state
+        log.message("[\(type(of: self))].\(#function) - \(state)")
+
+        let isEnabled = sender.state == .on ? true : false
+        presenter?.forceStatusMenusTwoLines(isEnabled)
+    }
+
+    @IBAction func comboBoxSecondLineDidChanged(_ sender: NSComboBox) {
+        let index = comboBoxStatusMenusUpdatePeriod.indexOfSelectedItem
+        log.message("[\(type(of: self))].\(#function) - \(index)")
+
+        presenter?.forceStatusMenusSecondLine(sender.indexOfSelectedItem)
+    }
+
+    @IBAction func comboBoxToolTipLeftDidChanged(_ sender: NSComboBox) {
+        let index = comboBoxStatusMenusUpdatePeriod.indexOfSelectedItem
+        log.message("[\(type(of: self))].\(#function) - \(index)")
+
+        presenter?.forceStatusMenusToolTipLeft(sender.indexOfSelectedItem)
+    }
+
+    @IBAction func comboBoxToolTipRightDidChanged(_ sender: NSComboBox) {
+        let index = comboBoxStatusMenusUpdatePeriod.indexOfSelectedItem
+        log.message("[\(type(of: self))].\(#function) - \(index)")
+
+        presenter?.forceStatusMenusToolTipRight(sender.indexOfSelectedItem)
     }
 
     // MARK: - OpenWeatherKey Input
@@ -224,11 +265,12 @@ extension OptionsViewController: OptionsViewDelegate {
     }
 
     func onViewDidAppear() {
-        refreshOptionsViewStates()
+        refreshUI()
     }
 
-    func refreshStatusMenusUpdatePeriod() {
+    func refreshStatusMenusOptions() {
         udpateComboBoxStatusMenusUpdatePeriod()
+        updateStatusMenusViewOptions()
     }
 
     // MARK: - MVPViewDelegate
@@ -239,13 +281,16 @@ extension OptionsViewController: OptionsViewDelegate {
 
         controlOpenWeatherKey.delegate = self
 
-        if #unavailable(macOS 10.14) {  // For HighSierra only.
+        if isHighSierra {
+
             boxAppOptions.isTransparent = true
             boxWeatherOptions.isTransparent = true
             boxSpecialOptions.isTransparent = true
 
-            // Dark Mode is .system (by default) only
-            // controlDarkMode.isEnabled = false
+        }
+
+        if isLegacy == false {
+            buttonInfo.isHidden = true
         }
 
         lockOpenWeatherKeyHole()
@@ -312,10 +357,6 @@ extension OptionsViewController: OptionsViewDelegate {
         boxSpecialOptions.title = "Section: Special Options".localizedValue + ":"
 
         labelOpenWeatherKey.stringValue = "Option: OpenWeather Key".localizedValue
-        labelStatusMenus.stringValue = "Option: StatusMenus".localizedValue
-
-        let weatherStatusMenusUpdate = "Option: StatusMenus Update".localizedValue + ":"
-        labelStatusMenusUpdate.stringValue = weatherStatusMenusUpdate
 
         controlOpenWeatherKey.placeholderString = controlOpenWeatherKey.isEditable ?
             "OpenWeather: Editable".localizedValue :
@@ -325,9 +366,22 @@ extension OptionsViewController: OptionsViewDelegate {
             "OpenWeather: Lock".localizedValue :
             "OpenWeather: Unlock".localizedValue
 
-        checkBoxStatusMenus.title = "Button: CheckBox StatusMenus".localizedValue
+        labelCurrentWeatherStatusMenus.stringValue =
+        "Option: CurrentWeatherStatusMenus".localizedValue
+
+        let weatherStatusMenusUpdate = "Option: StatusMenus Update".localizedValue + ":"
+        labelStatusMenusUpdate.stringValue = weatherStatusMenusUpdate
+        checkBoxCurrentWeatherStatusMenus.title = "Button: CheckBox StatusMenus".localizedValue
 
         udpateComboBoxStatusMenusUpdatePeriod()
+
+        labelStatusMenus.stringValue = "Option: StatusMenusView".localizedValue
+        checkBoxTwoLines.title = "Button: CheckBox TwoLines".localizedValue
+        labelSecondLine.stringValue = "Option: SecondLine".localizedValue + ":"
+        labelToolTip.stringValue = "Option: ToolTip".localizedValue + ":"
+
+        updateStatusMenusViewOptions()
+        buttonInfo.toolTip = "Button: TwoLines Info".localizedValue
 
         buttonClose.title = "Button: Close".localizedValue
         buttonResetAllSettings.title = "Button: Reset to Defaults".localizedValue
@@ -440,9 +494,9 @@ extension OptionsViewController {
         }
     }
 
-    private func udpateCheckBoxStatusMenus() {
+    private func udpateCheckBoxCurrentWeatherStatusMenus() {
         log.message("[\(type(of: self))].\(#function) \(AppOptions.statusMenusOption)")
-        checkBoxStatusMenus.state = AppOptions.statusMenusOption ? .on : .off
+        checkBoxCurrentWeatherStatusMenus.state = AppOptions.statusMenusOption ? .on : .off
     }
 
     private func udpateComboBoxStatusMenusUpdatePeriod() {
@@ -459,7 +513,33 @@ extension OptionsViewController {
         comboBoxStatusMenusUpdatePeriod.isEnabled = AppOptions.statusMenusOption
     }
 
-    private func refreshOptionsViewStates() {
+    private func updateStatusMenusViewOptions() {
+        log.message("[\(type(of: self))].\(#function)")
+        let options = AppOptions.statusMenusViewOptions
+
+        if isLegacy { // Legacy OS doesn't support multiline status menus items.
+            checkBoxTwoLines.isEnabled = false
+            comboBoxSecondLine.isEnabled = false
+        }
+
+        checkBoxTwoLines.state = options.twoLines ? .on : .off
+
+        comboBoxSecondLine.removeAllItems()
+        comboBoxToolTipLeft.removeAllItems()
+        comboBoxToolTipRight.removeAllItems()
+
+        for item in MeteoParameter.allCases {
+            comboBoxSecondLine.addItem(withObjectValue: "\(item)".localizedValue)
+            comboBoxToolTipLeft.addItem(withObjectValue: "\(item)".localizedValue)
+            comboBoxToolTipRight.addItem(withObjectValue: "\(item)".localizedValue)
+        }
+
+        comboBoxSecondLine.selectItem(at: options.secondLine.rawValue)
+        comboBoxToolTipLeft.selectItem(at: options.toolTipLeft.rawValue)
+        comboBoxToolTipRight.selectItem(at: options.toolTipRight.rawValue)
+    }
+
+    private func refreshUI() {
 
         log.message("[\(type(of: self))].\(#function)")
 
@@ -472,7 +552,8 @@ extension OptionsViewController {
         updateControlPressure()
         updateControlDistance()
 
-        udpateCheckBoxStatusMenus()
+        udpateCheckBoxCurrentWeatherStatusMenus()
         udpateComboBoxStatusMenusUpdatePeriod()
+        updateStatusMenusViewOptions()
     }
 }